First approach to lib
