+++ srcs/login/src/settings.py
@@ -56,6 +56,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'login.apps.LoginConfig',
     'corsheaders',
+    # 'tcommons'
 ]
 
 MIDDLEWARE = [
